Move dumper structures from sdk.h to clientclass.h
Rename GetFlags() -> Entity_GetFlags()
Typedef Hashmap
