+++ src/include/sdk.h
@@ -5,6 +5,8 @@
 #include <stdbool.h>
 #include <wchar.h>
 
+#include "netvars.h"
+
 #define STR(A, B) A##B
 #define PADSTR(N) STR(pad, N)
 #define PAD(N)    uint8_t PADSTR(__LINE__)[N]
@@ -129,69 +131,6 @@ typedef struct {
     bool m_bViewToProjectionOverride;
     VMatrix m_ViewToProjection;
 } ViewSetup;
-
-typedef enum {
-    INT = 0,
-    FLOAT,
-    VECTOR,
-    VECTOR2D,
-    STRING,
-    ARRAY,
-    DATATABLE,
-    INT64,
-    SENDPROPTYPEMAX
-} SendPropType;
-
-typedef struct DataVariant_s { /* struct unions 4ever */
-    union {
-        float Float;
-        long Int;
-        char* String;
-        void* Data;
-        float Vector[3];
-        int64_t Int64;
-    };
-
-    SendPropType type;
-} DataVariant;
-
-typedef struct RecvProp_s RecvProp;
-typedef struct RecvTable_s {
-    struct RecvProp_s* props;
-    int propsCount;
-    void* decoder;
-    char* tableName;
-    bool initialized;
-    bool inMainList;
-} RecvTable;
-
-typedef struct RecvProp_s {
-    char* varName;
-    SendPropType recvType;
-    int flags;
-    int stringBufferSize;
-    bool insideArray;
-    const void* extraData;
-    RecvProp* arrayProp;
-    void* arrayLengthProxyFn;
-    void* proxyFn;
-    void* dataTableProxyFn;
-    RecvTable* dataTable;
-    int offset;
-    int elementStride;
-    int elements;
-    const char* parentArrayPropName;
-} RecvProp;
-
-typedef struct ClientClass_s {
-    void* CreateClientClassFn;
-    void* CreateEventFn;
-    char* network_name;
-    RecvTable* recv_table;
-    struct ClientClass_s* next;
-    int class_id;
-    const char* map_class_name;
-} ClientClass;
 
 typedef struct {
     VectorAligned m_Start;
@@ -325,6 +264,7 @@ typedef struct {
 /* Classes */
 
 #include "sdk/studiohdr.h"
+#include "sdk/clientclass.h"
 #include "sdk/entity.h"
 #include "sdk/weapon.h"
 #include "sdk/convar.h"
@@ -662,7 +602,5 @@ void TraceHull(vec3_t start, vec3_t end, vec3_t hull_min, vec3_t hull_max,
                uint32_t mask, Trace_t* trace);
 
 bool IsBehindAndFacingTarget(Entity* owner, Entity* target);
-
-int GetFlags(Entity* ent);
 
 #endif /* SDK_H_ */
